+++ CHANGELOG.md
@@ -2,7 +2,15 @@
 
 ## v0.3.0 (in development)
 
-* Major refactor to work with `polars>=0.20.8`
+* Major refactor to work with `polars>=1.0.0`
+
+#### Functionality improvements
+
+* Convert `by` arg `_by` to allow naming columns `by` in `.mutate()`/`.summarize()`
+
+#### New functions
+
+* `where()`
 
 ## v0.2.19